feat(Exchange awareness): EXCXP-1963 - Add Exchange announcement. (#4951)
## Objective

[JIRA Ticket Link](https://blockchain.atlassian.net/browse/EXCXP-1963)

In order to be able to increase the Exchange awareness for our users, we
need to show a new announcement that will be only visible for users
within a given cohort.

## Description

I tried to mimic as much as possible all other announcements and the
Exchange services so I could have the following:

An exchange announcement that will be only visible if the new endpoint
we've deployed for that returns a cohort equal or greater than 1 and the
feature flags is on. That announcement needs to be recurrent using the
time obtained with the feature flag setup.

When the user taps on the announcement action, it will trigger a new
dynamic link (please update your local configs because the network
configs are already updated) that will either open the exchange app with
a UUID and the cohort; or open the app store. (That UUID and cohort will
be send as part of the analytics that will be implemented on top of this
work, and also sent from the Exchange once the link is open in that end
so our BI team can correlate them)

## How to Test

Update your local config, build the staging app and just open the app.
(If you don't see the announcement is because your user does not belong
to the desired cohort)

## Screenshot

[Figma
Link](https://www.figma.com/file/zmLgDsKEOgCISPqoFavsb6/iOS---Exchange-%2F-Q3-Q4-2022?node-id=1763%3A74232&t=9FJtx8QYE7TGf3xN-4)



https://user-images.githubusercontent.com/87308583/207938340-3598e7a8-821e-4c52-8bc0-3f43775b48c8.mov



## Merge Checklist

- [X] The PR uses a title supported by
[.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [X] You have added sufficient documentation (descriptive comments), if
applicable.