feat(Withdrawal Locks): IOS-7474 - On Hold on new Trading Dashboard (#5000)
## Objective

[JIRA Ticket Link](https://blockchain.atlassian.net/browse/IOS-7474)
On Hold on new Trading Dashboard

## Description

Add on hold component under assets in dashboard. Clicking takes you to
the current on hold view (live in prod) in a modal.

## How to Test

If you have some on hold balance, it will show up in dashboard. If you
dont, you may change `var showOnHoldSection: Bool` to always return true
to test it out.

## Screenshot

![Simulator Screen Shot - iPhone 14 - 2023-01-11 at 10 45
05](https://user-images.githubusercontent.com/60097845/211822008-4c028cb2-ab33-4076-86b0-9c08f6456a54.png)
![Simulator Screen Shot - iPhone 14 - 2023-01-11 at 10 45
28](https://user-images.githubusercontent.com/60097845/211822020-1f201d79-995b-4e94-b143-ca2645ce5770.png)


## Merge Checklist

- [X] The PR uses a title supported by
[.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [X] You have added sufficient documentation (descriptive comments), if
applicable.