feat(Brokerage): Add a cannon to the success screen (#5165)
## Description

Upon transaction success, show a confetti cannon - the user can tap the
screen again to see the cannon fire more confetti.

At the moment this relies on the old copy on the screen, however it is a
start to create a new space to improve the success screens. We should
look at controlling more of this via backend order api - there is no
reason why this should be hard-coded in client

## How to Test

Buy some crypto, feature flag is turned on currently. See:
```
blockchain.ux.transaction[buy].pending.transaction.is.enabled
```

## Screenshot


https://user-images.githubusercontent.com/88668746/220617272-2da947ca-2cf1-4068-aa14-861ffa8c039e.mp4

## Merge Checklist

- [x] The PR uses a title supported by
[.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).