fix(Send): fix asset loading when app starts in trading mode (#5758)
## Objective

fix asset loading when app starts in trading mode

## Description

* Filter your console with '🫂' emoji and check the logs. 
* When starting in DeFi mode, all erc20s will load.
* When starting in Trading mode, we load all custodial assets, when we
switch to DeFi mode we replace those assets with non custodial ones. In
this case we end up not loading any ERC20 assets for which the user has
balance for and are not custodially supported. This happens due to us
calling `ERC20CryptoAssetServiceAPI` at the wrong time (while we were in
custodial mode).
* This PRs changes how this is setup. `ERC20CryptoAssetServiceAPI` will
now inject a handler into `Coincore` so we control non custodial asset
loading from a single place, respecting the `AppMode`.

## How to Test

* Check AssetLoader working, filter your console with '🫂' emoji and
check the logs.
* Open the app in DeFi mode and check logs
* Open app in Trading mode, check logs, then switch and check all your
assets were loaded.
* The Send screen will now contain everything that was missing.

## Screenshot


https://github.com/blockchain/wallet-ios-private/assets/60097845/b2ea756d-9676-4a4b-9c21-14b0d9b5a775

## Merge Checklist

- [X] The PR uses a title supported by
[.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [ ] You have added sufficient documentation (descriptive comments), if
applicable.