fix(growth): GROWUX-0000 Change the default app mode to PKW (#6093)
## Objective

Since trading is unavailable in high risk countries such as Ghana and
Nigeria, it's dangerous to default the AppMode to trading.
The user can arrive in a state where the appMode is trading but the the
products endpoint dictate that the mode should be disabled.
This may be the source of our "Blank App" bugs reported by users from
those countries.

As a first step, we should default to PKW instead of trading. 
I will take a look at the whole AppMode handling in the AppChrome in the
next sprint.