fix(AppChrome): fix dismiss all (#6164)
## Objective

fix dismiss all so it doesn't dismiss app chrome.
This is reproducible in iOS 16.x

In our flows, we present stuff on top of AppChrome, that itself is a
presented sheet.

`SuperAppRootController > AppChrome > DEX Choose product sheet`

So whenever we want to dismiss the whole lot, we need to not dismiss
`AppChrome`.

The original code here was
```
func dismissAll(animated: Bool = true, completion: (() -> Void)? = nil) {
  guard let hostingController = presentedViewController else {
    return
  }
  guard let appChrome = hostingController.presentedViewController else {
    return
  }
  if appChrome.isBeingDismissed {
    app.post(error: NavigationError.isBeingDismissedError(appChrome))
  }
  appChrome.dismiss(animated: animated, completion: completion)
}
```
but this had a bug where we weren't dismissing the first screen past the
AppChrome, so we 'fixed' that and caused the bug that sometimes we would
dismiss the AppChrome itself.

Now, we fixed that and only dismiss the presented screen if there is
something show on top of appchrome. 🤞

## How to Test

1. Defi -> Swap -> Dex
2. Defi -> USDC (or any other currency supported by Swap and Dex) -> Dex

## Screenshot


https://github.com/blockchain/wallet-ios-private/assets/60097845/f6349cd5-7d92-4481-9f8a-dfb69997cb8e

## Merge Checklist

- [X] The PR uses a title supported by
[.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [ ] You have added sufficient documentation (descriptive comments), if
applicable.