fix(custodial activity): growux-2165 -  hit /payments/transactions only once per product (#6187)
## Objective

Today, we hit /payments/transactions only once per trading currency per
product, plus once for the activity of the active trading currency.
In this PR we change that so we will hit it once per product only, and
slightly reduced the cache expiration.

## Description

We now hit the endpoint 4 times, instead of 68. 

## How to Test

Open app
- check dashboard loads
- check all activity screen loads
- check activity details loads

## Merge Checklist

- [X] The PR uses a title supported by
[.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [ ] You have added sufficient documentation (descriptive comments), if
applicable.