fix(growth): GROWUX-2249 Swap fixes & improvements (#6207)
## Objective

Swap was giving all sorts of problems because of how limits were
handled.
I've refactored the view to take into account errors as well as limits. 
Now every state should be reflected in the UI.

Note: 
- there seems to be an error in the transaction engine when you input a
small value. There is no error being returned even if the input is
clearly under the min spendable amount.
To work around this issue I am also using the limits value as a
safeguard.
Lastly, there seemed to be an error with the string for limits. Because
of the way the mint amount was being rounded you could have situations
in which the title and message contain two different min values (8.14€ &
9€)

[JIRA Ticket Link](https://blockchain.atlassian.net/browse/GROWUX-2249)



https://github.com/blockchain/wallet-ios-private/assets/102520557/52194d00-9bb6-4280-8ac7-d2b7bf2761af